bugfix(#116): Fix dependencies
