+++ backend/services/calendar/package.json
@@ -30,7 +30,7 @@
     "@graphql-codegen/typescript": "^1.17.11",
     "@graphql-codegen/typescript-resolvers": "^1.17.11",
     "@types/chai": "^4.2.14",
-    "@types/chai-spies": "^1.0.2",
+    "@types/chai-spies": "^1.0.3",
     "@types/mocha": "^8.0.4",
     "@types/mock-knex": "^0.4.3",
     "apollo-server-testing": "^2.19.0",
