+++ backend/services/calendar/src/index.ts
@@ -1,28 +1,6 @@
-import { readFileSync } from 'fs';
-import { resolve } from 'path';
-import { ApolloServer, gql } from 'apollo-server';
-import { buildFederatedSchema } from '@apollo/federation';
-
+import createApolloServer from './server';
 import { context } from 'dsek-shared';
-
 import dataSources from './datasources';
-
-const typesSrc = readFileSync(resolve(__dirname, 'schema.graphql'));
-const typeDefs = gql`${typesSrc}`
-import resolvers from './resolvers';
-
-export const createApolloServer = (context: any, dataSources?: any) => {
-  return new ApolloServer({
-    schema: buildFederatedSchema([
-      {
-        typeDefs,
-        resolvers
-      }
-    ]),
-    context: context,
-    dataSources: dataSources,
-  });
-}
 
 const server = createApolloServer(context.deserializeContext, dataSources);
